Move download code to download-utils crate
