+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
     "cli-config",
     "client",
     "core",
+    "download-utils",
     "faucet",
     "perf",
     "validator",
