+++ download-utils/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "solana-download-utils"
+version = "1.1.0"
+description = "Solana Download Utils"
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+edition = "2018"
+
+[dependencies]
+bzip2 = "0.3.3"
+console = "0.9.2"
+indicatif = "0.14.0"
+log = "0.4.8"
+reqwest = { version = "0.10.1", default-features = false, features = ["blocking", "rustls-tls", "json"] }
+solana-sdk = { path = "../sdk", version = "1.1.0" }
+solana-ledger = { path = "../ledger", version = "1.1.0" }
+tar = "0.4.26"
+
+[lib]
+crate-type = ["lib"]
+name = "solana_download_utils"
