+++ local-cluster/Cargo.toml
@@ -16,6 +16,7 @@ solana-archiver-lib = { path = "../archiver-lib", version = "1.1.0" }
 solana-config-program = { path = "../programs/config", version = "1.1.0" }
 solana-core = { path = "../core", version = "1.1.0" }
 solana-client = { path = "../client", version = "1.1.0" }
+solana-download-utils = { path = "../download-utils", version = "1.1.0" }
 solana-faucet = { path = "../faucet", version = "1.1.0" }
 solana-exchange-program = { path = "../programs/exchange", version = "1.1.0" }
 solana-genesis-programs = { path = "../genesis-programs", version = "1.1.0" }
