+++ validator/Cargo.toml
@@ -9,18 +9,16 @@ license = "Apache-2.0"
 homepage = "https://solana.com/"
 
 [dependencies]
-bzip2 = "0.3.3"
 clap = "2.33.0"
 chrono = { version = "0.4.10", features = ["serde"] }
 console = "0.9.2"
 log = "0.4.8"
-indicatif = "0.14.0"
 rand = "0.6.5"
-reqwest = { version = "0.10.1", default-features = false, features = ["blocking", "rustls-tls", "json"] }
 serde_json = "1.0.46"
 solana-clap-utils = { path = "../clap-utils", version = "1.1.0" }
 solana-client = { path = "../client", version = "1.1.0" }
 solana-core = { path = "../core", version = "1.1.0" }
+solana-download-utils = { path = "../download-utils", version = "1.1.0" }
 solana-faucet = { path = "../faucet", version = "1.1.0" }
 solana-ledger = { path = "../ledger", version = "1.1.0" }
 solana-logger = { path = "../logger", version = "1.1.0" }
@@ -31,7 +29,6 @@ solana-runtime = { path = "../runtime", version = "1.1.0" }
 solana-sdk = { path = "../sdk", version = "1.1.0" }
 solana-vote-program = { path = "../programs/vote", version = "1.1.0" }
 solana-vote-signer = { path = "../vote-signer", version = "1.1.0" }
-tar = "0.4.26"
 
 [target."cfg(unix)".dependencies]
 gag = "0.1.10"
